# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface YoutubeSearchResult {
  videoId: string;
  title: string;
  thumbnail: string;
  duration: string;
  channel: string;
  views: string;
}

export interface YoutubeSearchResponse {
  results: YoutubeSearchResult[];
}

export type SearchYoutubeParams = {
/**
 * @minLength 1
 */
q: string;
};

export type DownloadYoutubeParams = {
/**
 * @minLength 1
 */
videoId: string;
format: DownloadYoutubeFormat;
};

export type DownloadYoutubeFormat = typeof DownloadYoutubeFormat[keyof typeof DownloadYoutubeFormat];


export const DownloadYoutubeFormat = {
  mp3: 'mp3',
  mp4: 'mp4',
} as const;