# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: youtube
    description: YouTube search and download operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /youtube/search:
    get:
      operationId: searchYoutube
      tags: [youtube]
      summary: Search YouTube videos
      description: Searches YouTube for videos matching a query
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 1
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/YoutubeSearchResponse"
        "400":
          description: Invalid query
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /youtube/download:
    get:
      operationId: downloadYoutube
      tags: [youtube]
      summary: Download a YouTube video as MP4 or MP3
      description: Streams the requested video's audio or video track as a downloadable file
      parameters:
        - name: videoId
          in: query
          required: true
          schema:
            type: string
            minLength: 1
        - name: format
          in: query
          required: true
          schema:
            type: string
            enum: [mp3, mp4]
      responses:
        "200":
          description: File stream
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: Upstream YouTube error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    YoutubeSearchResult:
      type: object
      properties:
        videoId:
          type: string
        title:
          type: string
        thumbnail:
          type: string
        duration:
          type: string
        channel:
          type: string
        views:
          type: string
      required:
        - videoId
        - title
        - thumbnail
        - duration
        - channel
        - views
    YoutubeSearchResponse:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: "#/components/schemas/YoutubeSearchResult"
      required:
        - results