# lib/api-zod/src/generated/types/youtubeSearchResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { YoutubeSearchResult } from './youtubeSearchResult';

export interface YoutubeSearchResponse {
  results: YoutubeSearchResult[];
}