# lib/api-zod/src/generated/types/youtubeSearchResult.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface YoutubeSearchResult {
  videoId: string;
  title: string;
  thumbnail: string;
  duration: string;
  channel: string;
  views: string;
}